Allowed year to be passed into getData function
