+++ index.js
@@ -1,5 +1,3 @@
 import getData from './getData.js';
-import getDataNodeList from './getDataNodeList.js';
 
-/* getData(); */
-getDataNodeList();
+getData(2010);
